SP- 42 Refactor arrays to lists
Refactored arrays to list in BubbleSort.java and updated test files PrintArray.java and Test.java accordingly. Test.java updated to create list with random numbers. Tests pass in separate environment. I was unable to run in this environment because files are not on class path and did not want to update the build.gradle.kts dependencies without checking with the team first. Note that Tests.java should also be moved to the test/java directory, if it is still necessary to keep file after this step.