SP-36 Add Toggle Ascending or Descending
SP-37-Refactor-bubble-sort-algorithm-files
Update the algorithm, it was not working properly.
We are now able to change from ascending to descending order.
Application does not have error pop up, but instead auto generates a range and numbers when a user presses the submit button.
Created a Swap class to store our data when swapping the numbers.
Added more functionality to BubbleSort class to store data and return the string to display it the user.